better utility creation
