+++ pkg/utils/fileutils/compress.go
@@ -1,0 +1,55 @@
+package fileutils
+
+import (
+	"archive/tar"
+	"compress/gzip"
+	"io"
+	"os"
+	"path/filepath"
+	"strings"
+)
+
+// CreateTarFile creates a gzipped tar file from srcDir and writes it to outPath.
+func CreateTarFile(srcDir, outPath string) error {
+	mw, err := os.Create(outPath)
+	if err != nil {
+		return err
+	}
+	defer mw.Close()
+	gzw := gzip.NewWriter(mw)
+	defer gzw.Close()
+
+	tw := tar.NewWriter(gzw)
+	defer tw.Close()
+
+	return filepath.Walk(srcDir, func(file string, fi os.FileInfo, err error) error {
+		if err != nil {
+			return err
+		}
+		if !fi.Mode().IsRegular() {
+			return nil
+		}
+		header, err := tar.FileInfoHeader(fi, fi.Name())
+		if err != nil {
+			return err
+		}
+		header.Name = strings.TrimPrefix(strings.Replace(file, srcDir, "", -1), string(filepath.Separator))
+		header.Size = fi.Size()
+		header.Mode = int64(fi.Mode())
+		header.ModTime = fi.ModTime()
+		if err := tw.WriteHeader(header); err != nil {
+			return err
+		}
+
+		f, err := os.Open(file)
+		if err != nil {
+			return err
+		}
+		defer f.Close()
+
+		if _, err := io.Copy(tw, f); err != nil {
+			return err
+		}
+		return nil
+	})
+}
